Move RecyclerView ViewHolders out of common They're only used in mobile module
