+++ mobile/src/main/java/com/boswelja/devicemanager/common/ui/IconOneLineViewHolder.kt
@@ -5,13 +5,13 @@
  * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
  * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
  */
-package com.boswelja.devicemanager.common.recyclerview.item
+package com.boswelja.devicemanager.common.ui
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.annotation.DrawableRes
 import androidx.recyclerview.widget.RecyclerView
-import com.boswelja.devicemanager.common.databinding.RecyclerviewItemIconOneLineBinding
+import com.boswelja.devicemanager.databinding.RecyclerviewItemIconOneLineBinding
 
 open class IconOneLineViewHolder(private val binding: RecyclerviewItemIconOneLineBinding) :
     RecyclerView.ViewHolder(binding.root) {
