+++ mobile/src/main/java/com/boswelja/devicemanager/common/ui/OneLineViewHolder.kt
@@ -5,12 +5,12 @@
  * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
  * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
  */
-package com.boswelja.devicemanager.common.recyclerview.item
+package com.boswelja.devicemanager.common.ui
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.boswelja.devicemanager.common.databinding.RecyclerviewItemOneLineBinding
+import com.boswelja.devicemanager.databinding.RecyclerviewItemOneLineBinding
 
 open class OneLineViewHolder private constructor(
     private val binding: RecyclerviewItemOneLineBinding
