+++ mobile/src/main/java/com/boswelja/devicemanager/dndsync/ui/helper/SetupFragment.kt
@@ -15,7 +15,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.boswelja.devicemanager.R
-import com.boswelja.devicemanager.common.recyclerview.adapter.StringAdapter
+import com.boswelja.devicemanager.common.ui.adapter.StringAdapter
 import com.boswelja.devicemanager.databinding.FragmentDndsyncHelperSetupBinding
 
 internal class SetupFragment : Fragment() {
