+++ mobile/src/main/res/layout/recyclerview_item_icon_one_line.xml
