+++ mobile/src/main/res/layout/recyclerview_item_one_line.xml
